requirements.txt
click==8.1.7
colorama==0.4.6
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.0.1
pillow==10.4.0
